feat:organize and add new service funcion
