+++ gammerRunner-tutorial/src/main/java/jordan/tutorials/gammerRunnertutorial/Enterprise/Data/DataService.java
@@ -1,4 +1,4 @@
-package jordan.tutorials.gammerRunnertutorial.Bussiness.example;
+package jordan.tutorials.gammerRunnertutorial.Enterprise.Data;
 
 import org.springframework.stereotype.Component;
 
@@ -10,4 +10,5 @@ public class DataService {
     public List<Integer> getData(){
         return Arrays.asList(10,20,30,40);
     }
+    public List<String> getDataList(){return Arrays.asList("Hola", "Jordan", "comming soon");}
 }
