+++ gammerRunner-tutorial/src/main/java/jordan/tutorials/gammerRunnertutorial/Enterprise/Web/MyWebController.java
@@ -1,5 +1,6 @@
-package jordan.tutorials.gammerRunnertutorial.Bussiness.example;
+package jordan.tutorials.gammerRunnertutorial.Enterprise.Web;
 
+import jordan.tutorials.gammerRunnertutorial.Enterprise.Bussiness.BussinessLogic;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -14,5 +15,8 @@ public class MyWebController {
 
     public long returnValueFromBussinesServuce(){
         return bussinessLogic.calculateSum();
+    }
+    public String returnValueStringFromBussinesSservice(){
+        return bussinessLogic.arrayJoin();
     }
 }
