+++ gammerRunner-tutorial/src/main/java/jordan/tutorials/gammerRunnertutorial/GammerRunnerTutorialApplication.java
@@ -1,10 +1,7 @@
 package jordan.tutorials.gammerRunnertutorial;
 
-import jordan.tutorials.gammerRunnertutorial.Bussiness.example.MyWebController;
+import jordan.tutorials.gammerRunnertutorial.Enterprise.Web.MyWebController;
 import jordan.tutorials.gammerRunnertutorial.gamer.GameRunner;
-import jordan.tutorials.gammerRunnertutorial.gamer.Stucture.MarioGame;
-import jordan.tutorials.gammerRunnertutorial.gamer.Stucture.PackmanGame;
-import jordan.tutorials.gammerRunnertutorial.gamer.Stucture.SuperContractGame;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -18,6 +15,7 @@ public class GammerRunnerTutorialApplication {
 		GameRunner runner = context.getBean(GameRunner.class);
 		MyWebController controller = context.getBean(MyWebController.class);
 		System.out.println(controller.returnValueFromBussinesServuce());
+		System.out.println(controller.returnValueStringFromBussinesSservice());
 		runner.run();
 	}
