+++ gammerRunner-tutorial/src/main/java/jordan/tutorials/gammerRunnertutorial/gamer/Stucture/PackmanGame.java
@@ -3,6 +3,7 @@ package jordan.tutorials.gammerRunnertutorial.gamer.Stucture;
 import jordan.tutorials.gammerRunnertutorial.gamer.interfaces.GammingConsole;
 import org.springframework.stereotype.Component;
 
+@Component
 public class PackmanGame  implements GammingConsole {
     public PackmanGame(){};
