+++ gammerRunner-tutorial/src/main/java/jordan/tutorials/gammerRunnertutorial/gamer/Stucture/SuperContractGame.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 @Component
-@Primary
 public class SuperContractGame implements GammingConsole {
     public SuperContractGame(){}
     public void up(){
